requirements.txt
streamlit==1.25.0  # Replace with your current version
qrcode==7.4.2      # QR code generation
Pillow==10.0.1     # Image manipulation for QR code logo
